Scan logs on change in current query
