+++ extensions/ql-vscode/src/extension.ts
@@ -101,6 +101,7 @@ import { RemoteQuery } from './remote-queries/remote-query';
 import { EvalLogViewer } from './eval-log-viewer';
 import { SummaryLanguageSupport } from './log-insights/summary-language-support';
 import { JoinOrderScannerProvider } from './log-insights/join-order';
+import { LogScannerService } from './log-insights/log-scanner-service';
 
 /**
  * extension.ts
@@ -485,7 +486,9 @@ async function activateWithInstalledDistribution(
   ctx.subscriptions.push(qhm);
 
   void logger.log('Initializing evaluation log scanners.');
-  ctx.subscriptions.push(qhm.registerLogScannerProvider(new JoinOrderScannerProvider()));
+  const logScannerService = new LogScannerService(qhm);
+  ctx.subscriptions.push(logScannerService);
+  ctx.subscriptions.push(logScannerService.registerLogScannerProvider(new JoinOrderScannerProvider()));
 
   void logger.log('Reading query history');
   await qhm.readQueryHistory();
@@ -520,8 +523,6 @@ async function activateWithInstalledDistribution(
     forceReveal: WebviewReveal
   ): Promise<void> {
     await intm.showResults(query, forceReveal, false);
-    // Always update the log warnings so they stay in sync with the results.
-    await qhm.scanEvalLog(query);
   }
 
   async function compileAndRunQuery(
@@ -562,7 +563,7 @@ async function activateWithInstalledDistribution(
           undefined,
           item,
         );
-        item.completeThisQuery(completedQueryInfo);
+        qhm.completeQuery(item, completedQueryInfo);
         await showResultsForCompletedQuery(item as CompletedLocalQueryInfo, WebviewReveal.NotForced);
         // Note we must update the query history view after showing results as the
         // display and sorting might depend on the number of results
